refactor: move children project request from ProjectListItemConnected to ProjectPage
